+++ api/server.js
@@ -86,6 +86,13 @@ if ((!process.env.KEYSTORE || !process.env.KEYSTORE_BUCKET) && config.env === 'p
           for (let i = 0; i < server.connections.length; i++) {
             const message = `${server.connections[i].info.protocol === 'https' ? 'SSL ' : ''}Server running at: ${server.connections[i].info.uri}`;
             server.log(['info'], message);
+
+            const keepAliveTimeout = process.env.KEEP_ALIVE_TIMEOUT || 60;
+  
+            // keep tcp connections open a bit longer than the load balancer's timeout to prevent hangups that cause 502s
+            server.connections[i].listener.keepAliveTimeout = (keepAliveTimeout * 1000);
+            // headersTimeout timer starts _after_ keep-alive timer has started so give it a 1 second buffer to prevent hangups
+            server.connections[i].listener.headersTimeout = (keepAliveTimeout * 1000) + 1000;
           }
         });
       } catch (e) {