+++ api/src/lib/logger.js
@@ -1,23 +1,23 @@
 const RollbarTransport = require('winston-rollbar-transport').default;
-const Sentry = require('winston-raven-sentry');
 const winston = require('winston');
-const transports = [new winston.transports.Console()];
+const SentryTransport = require('@synapsestudios/winston-sentry');
 const config = require('../../config');
+const transports = [new winston.transports.Console()];
 
-if (process.env.SENTRY_DSN) {
-  transports.push(new Sentry({
-    dsn: process.env.SENTRY_DSN,
-    level: 'error',
-    config: {
-      environment: config('/env'),
-    },
-  }));
+if (config('/errorLogging/sentryDSN')) {
+  const Sentry = require('./sentry');
+  transports.push(
+    new SentryTransport({
+      Sentry,
+      // Filter out timestamp so errors are grouped together better
+      formatter: (options) => options.message.replace(/\d{6}\/\d{6}\.\d{3},\s\[.*\]\smessage:\s(.*),\s.*/, '$1')
+    })
+  )
 }
 
-
-if (process.env.ROLLBAR_ACCESS_TOKEN) {
+if (config('/errorLogging/rollbarAccessToken')) {
   transports.push(new RollbarTransport({
-    rollbarAccessToken: process.env.ROLLBAR_ACCESS_TOKEN,
+    rollbarAccessToken: config('/errorLogging/rollbarAccessToken'),
     level: 'error',
   }));
 }