+++ api/src/lib/s3.js
@@ -1,0 +1,3 @@
+const aws = require('aws-sdk');
+
+module.exports = new aws.S3({ apiVersion: '2006-03-01' });