+++ api/src/plugins/openid-connect/views.js
@@ -2,13 +2,17 @@ const querystring = require('querystring');
 
 module.exports = {
   login: (cookie, client, options, error, login) => {
-    const queryString = querystring.stringify({
+    const queryStringParams = {
       client_id: cookie.params.client_id,
       response_type: cookie.params.response_type,
       scope: cookie.params.scope,
       redirect_uri: cookie.params.redirect_uri,
-      nonce: cookie.params.nonce
-    });
+      nonce: cookie.params.nonce,
+    };
+    if (cookie.params.code_challenge) {
+      queryStringParams.code_challenge = cookie.params.code_challenge;
+    }
+    const queryString = querystring.stringify(queryStringParams);
 
     return {
       client,
@@ -19,7 +23,7 @@ module.exports = {
       userRegistrationPath: `/user/register?${queryString}&login=true`,
       error,
       login,
-    }
+    };
   },
 
   interaction: (cookie, client) => ({