+++ api/test/lib/sentry-testkit.js
@@ -1,0 +1,3 @@
+const sentryTestkit = require('sentry-testkit')
+
+module.exports = sentryTestkit();